+++ deeppavlov/core/common/metrics_registry.json
@@ -1,5 +1,6 @@
 {
   "accuracy": "deeppavlov.metrics.accuracy:accuracy",
+  "acc": "deeppavlov.metrics.accuracy:round_accuracy",
   "bleu": "deeppavlov.metrics.bleu:bleu",
   "bleu_advanced": "deeppavlov.metrics.bleu:bleu_advanced",
   "classification_accuracy": "deeppavlov.metrics.accuracy:classification_accuracy",
@@ -8,7 +9,9 @@
   "classification_log_loss": "deeppavlov.metrics.log_loss:classification_log_loss",
   "classification_roc_auc": "deeppavlov.metrics.roc_auc_score:classification_roc_auc_score",
   "exact_match": "deeppavlov.metrics.squad_metrics:exact_match",
+  "f1": "deeppavlov.metrics.fmeasure:round_f1",
   "google_bleu": "deeppavlov.metrics.bleu:google_bleu",
+  "log_loss": "deeppavlov.metrics.log_loss:sk_log_loss",
   "loss": "deeppavlov.models.ranking.metrics:triplet_loss",
   "ner_f1": "deeppavlov.metrics.fmeasure:ner_f1",
   "per_item_accuracy": "deeppavlov.metrics.accuracy:per_item_accuracy",
@@ -16,16 +19,13 @@
   "per_item_dialog_accuracy": "deeppavlov.metrics.accuracy:per_item_dialog_accuracy",
   "per_item_dialog_bleu": "deeppavlov.metrics.bleu:per_item_dialog_bleu",
   "per_token_accuracy": "deeppavlov.metrics.accuracy:per_token_accuracy",
-  "r@1": "deeppavlov.models.ranking.metrics:r_at_1",
-  "r@2": "deeppavlov.models.ranking.metrics:r_at_2",
-  "r@5": "deeppavlov.models.ranking.metrics:r_at_5",
-  "r@10": "deeppavlov.models.ranking.metrics:r_at_10",
+  "r@1": "deeppavlov.metrics.recall_at_k:r_at_1",
+  "r@2": "deeppavlov.metrics.recall_at_k:r_at_2",
+  "r@5": "deeppavlov.metrics.recall_at_k:r_at_5",
+  "r@10": "deeppavlov.metrics.recall_at_k:r_at_10",
   "r@1_insQA": "deeppavlov.models.ranking.metrics:r_at_1_insQA",
   "rank_response": "deeppavlov.models.ranking.metrics:rank_response",
   "sets_accuracy": "deeppavlov.metrics.accuracy:sets_accuracy",
   "slots_accuracy": "deeppavlov.metrics.accuracy:slots_accuracy",
-  "squad_f1": "deeppavlov.metrics.squad_metrics:squad_f1",
-  "s_acc": "deeppavlov.models.ranking.metrics:siamese_acc",
-  "s_f1": "deeppavlov.models.ranking.metrics:siamese_f1",
-  "s_log_loss": "deeppavlov.models.ranking.metrics:siamese_log_loss"
+  "squad_f1": "deeppavlov.metrics.squad_metrics:squad_f1"
 }
