+++ deeppavlov/models/ranking/metrics.py
@@ -13,38 +13,10 @@
 # limitations under the License.
 
 import numpy as np
-from sklearn.metrics import f1_score, log_loss
 
 from deeppavlov.core.common.metrics_registry import register_metric
 
 
-@register_metric('r@1')
-def r_at_1(y_true, y_pred):
-    return recall_at_k(y_true, y_pred, k=1)
-
-
-@register_metric('r@2')
-def r_at_2(y_true, y_pred):
-    return recall_at_k(y_true, y_pred, k=2)
-
-
-@register_metric('r@5')
-def r_at_5(labels, predictions):
-    return recall_at_k(labels, predictions, k=5)
-
-@register_metric('r@10')
-def r_at_10(labels, predictions):
-    return recall_at_k(labels, predictions, k=10)
-
-def recall_at_k(y_true, y_pred, k):
-    num_examples = float(len(y_pred))
-    predictions = np.array(y_pred)
-    predictions = np.flip(np.argsort(predictions, -1), -1)[:, :k]
-    num_correct = 0
-    for el in predictions:
-        if 0 in el:
-            num_correct += 1
-    return float(num_correct) / num_examples
 
 @register_metric('rank_response')
 def rank_response(y_true, y_pred):
@@ -74,49 +46,3 @@ def recall_at_k_insQA(y_true, y_pred, k):
                 flags[i][j] = 1.
     return np.mean((np.sum(flags, -1) >= 1.).astype(float))
 
-@register_metric('s_acc')
-def siamese_acc(y_true, y_predicted):
-    """
-    Calculate accuracy in terms of absolute coincidence
-
-    Args:
-        y_true: array of true values
-        y_predicted: array of predicted values
-
-    Returns:
-        portion of absolutely coincidental samples
-    """
-    predictions = list(map(lambda x: round(x), list(np.squeeze(y_predicted, 1))))
-    examples_len = len(y_true)
-    correct = sum([y1 == y2 for y1, y2 in zip(y_true, predictions)])
-    return correct / examples_len if examples_len else 0
-
-@register_metric('s_f1')
-def siamese_f1(y_true, y_predicted):
-    """
-    Calculate accuracy in terms of absolute coincidence
-
-    Args:
-        y_true: array of true values
-        y_predicted: array of predicted values
-
-    Returns:
-        F1 score of the positive class in binary classification
-    """
-    predictions = list(map(lambda x: round(x), list(np.squeeze(y_predicted, 1))))
-    return f1_score(y_true, predictions)
-
-@register_metric('s_log_loss')
-def siamese_log_loss(y_true, y_predicted):
-    """
-    Calculate accuracy in terms of absolute coincidence
-
-    Args:
-        y_true: array of true values
-        y_predicted: array of predicted values
-
-    Returns:
-        Log loss in binary classification
-    """
-    predictions = list(np.squeeze(y_predicted, 1))
-    return log_loss(y_true, predictions)
