+++ tests/test_configs/ranking/paraphrase_ident_paraphraser_test.json
@@ -80,8 +80,8 @@
     "epochs": 200,
     "batch_size": 256,
     "pytest_max_batches": 2,
-    "train_metrics": ["s_f1", "s_acc", "s_log_loss"],
-    "metrics": ["s_f1", "s_acc", "s_log_loss"],
+    "train_metrics": ["f1", "acc", "log_loss"],
+    "metrics": ["f1", "acc", "log_loss"],
     "validation_patience": 10,
     "val_every_n_epochs": 5,
     "log_every_n_batches": 12
