+++ tests/test_configs/ranking/paraphrase_ident_qqp_test.json
@@ -81,8 +81,8 @@
     "batch_size": 256,
     "pytest_max_batches": 2,
     "metric_optimization": "minimize",
-    "train_metrics": ["s_log_loss", "s_acc", "s_f1"],
-    "metrics": ["s_log_loss", "s_acc", "s_f1"],
+    "train_metrics": ["log_loss", "acc", "f1"],
+    "metrics": ["log_loss", "acc", "f1"],
     "validation_patience": 10,
     "val_every_n_epochs": 1,
     "log_every_n_batches": 525
